WriteOff/requirements.txt
appdirs==1.4.3
configparser==3.5.0
Django==1.10.5
django-betterforms==1.1.4
django-extensions==1.7.7
django-widget-tweaks==1.4.1
enum34==1.1.6
flake8==3.3.0
graphviz==0.6
mccabe==0.6.1
packaging==16.8
pkg-resources==0.0.0
pycodestyle==2.3.1
pyflakes==1.5.0
pygraphviz==1.3.1
pyparsing==2.2.0
six==1.10.0
